src/app/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
loguru==0.7.3
email-validator==2.3.0
pyjwt==2.10.1
passlib==1.7.4
asyncpg==0.30.0
bcrypt==4.3.0
smsaero-api-async==3.1.0